+++ MetarSharp/Methods/Convert/Speed/FromKnots.cs
@@ -14,7 +14,7 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Miles Per Hour as a decimal with set decimal places</returns>
-        public static decimal ToMilesPerHour(this double value, byte decimalPlaces) => Math.Round((decimal)value * (decimal)1.151, decimalPlaces);
+        public static decimal ToMilesPerHour(this double value, int decimalPlaces) => Math.Round((decimal)value * (decimal)1.151, decimalPlaces);
 
         /// <summary>
         /// Converts the given value from Knots to Meters Per Second
@@ -28,7 +28,7 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Meters Per Second as a decimal with set decimal places</returns>
-        public static decimal ToMetersPerSecond(this double value, byte decimalPlaces) => Math.Round((decimal)value / (decimal)1.944, decimalPlaces);
+        public static decimal ToMetersPerSecond(this double value, int decimalPlaces) => Math.Round((decimal)value / (decimal)1.944, decimalPlaces);
 
         /// <summary>
         /// Converts the given value from Knots to Kilometers Per Hour
@@ -42,6 +42,6 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Kilometers Per Hour as a decimal with set decimal places</returns>
-        public static decimal ToKilometersPerHour(this double value, byte decimalPlaces) => Math.Round((decimal)value * (decimal)1.852, decimalPlaces);
+        public static decimal ToKilometersPerHour(this double value, int decimalPlaces) => Math.Round((decimal)value * (decimal)1.852, decimalPlaces);
     }
 }
