+++ MetarSharp/Methods/Convert/Speed/FromMetersPerSecond.cs
@@ -14,7 +14,7 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Miles Per Hour as a decimal with set decimal places</returns>
-        public static decimal ToMilesPerHour(this double value, byte decimalPlaces) => Math.Round((decimal)value * (decimal)2.237, decimalPlaces);
+        public static decimal ToMilesPerHour(this double value, int decimalPlaces) => Math.Round((decimal)value * (decimal)2.237, decimalPlaces);
 
         /// <summary>
         /// Converts the given value from Meters Per Second to Kilometers Per Hour
@@ -28,7 +28,7 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Kilometers Per Hour as a decimal with set decimal places</returns>
-        public static decimal ToKilometersPerHour(this double value, byte decimalPlaces) => Math.Round((decimal)value * (decimal)3.6, decimalPlaces);
+        public static decimal ToKilometersPerHour(this double value, int decimalPlaces) => Math.Round((decimal)value * (decimal)3.6, decimalPlaces);
 
         /// <summary>
         /// Converts the given value from Meters Per Second to Knots
@@ -42,6 +42,6 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Knots as a decimal with set decimal places</returns>
-        public static decimal ToKnots(this double value, byte decimalPlaces) => Math.Round((decimal)value * (decimal)1.944, decimalPlaces);
+        public static decimal ToKnots(this double value, int decimalPlaces) => Math.Round((decimal)value * (decimal)1.944, decimalPlaces);
     }
 }
