+++ MetarSharp/Methods/Convert/Temperature/FromFahrenheit.cs
@@ -14,7 +14,7 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Celsius as a decimal with set decimal places</returns>
-        public static decimal ToCelsius(this double value, byte decimalPlaces) => Math.Round(((decimal)value - 32) * 5 / 9, decimalPlaces);
+        public static decimal ToCelsius(this double value, int decimalPlaces) => Math.Round(((decimal)value - 32) * 5 / 9, decimalPlaces);
 
         /// <summary>
         /// Converts the given value from Fahrenheit to Kelvin
@@ -28,6 +28,6 @@
         /// </summary>
         /// <param name="value"></param>
         /// <returns>the value converted to Kelvin as a decimal with set decimal places</returns>
-        public static decimal ToKelvin(this double value, byte decimalPlaces) => Math.Round(((decimal)value - 32) * 5 / 9 + (decimal)273.15, decimalPlaces);
+        public static decimal ToKelvin(this double value, int decimalPlaces) => Math.Round(((decimal)value - 32) * 5 / 9 + (decimal)273.15, decimalPlaces);
     }
 }
